+++ alinkjs/packages/assetlink-pwa/vue.config.js
@@ -136,7 +136,7 @@ module.exports = defineConfig({
   configureWebpack: (config) => {
     config.resolve.alias = {
       ...config.resolve.alias,
-      'quasar': path.resolve('./node_modules/quasar'),
+      quasar: path.resolve("./node_modules/quasar"),
     };
 
     config.resolve.fallback = {