+++ alinkjs/packages/assetlink-sidecar/vue.config.js
@@ -100,7 +100,7 @@ module.exports = defineConfig({
   configureWebpack(config) {
     config.resolve.alias = {
       ...config.resolve.alias,
-      'quasar': path.resolve('./node_modules/quasar'),
+      quasar: path.resolve("./node_modules/quasar"),
     };
 
     config.plugins.push(